python_models/requirements.txt
# Smart Parking Python Models Requirements
# Number Plate Recognition & Weather Recommendation

# Core ML/DL
tensorflow>=2.15.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Number Plate Recognition
easyocr>=1.7.0
pytesseract>=0.3.10
ultralytics>=8.0.0  # YOLOv8 for object detection

# Image Processing
Pillow>=10.0.0
imutils>=0.5.4

# Weather API & Recommendation
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Firebase Integration
firebase-admin>=6.2.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
